refactor: kebab-case filenames
magic-kebab, run against this repo's own oxlint config: 19 renames and 31
import specifiers rewritten, no conflicts and nothing skipped. The three
NEEDS REVIEW hits were all display strings in src/i18n/locales/en-US.json
that happen to read "Awards" and "Publications" — prose, not specifiers, so
correctly left alone.

Renames only, so `git log --follow` has the easiest possible job inferring
them.
