+++ src/components/portfolio/shared.tsx
