+++ src/components/portfolio/writing-list.tsx
@@ -1,7 +1,7 @@
 import Balancer from "react-wrap-balancer";
 import { getTranslations } from "next-intl/server";
 import { MEDIUM, WRITING_TOPICS } from "./data";
-import { ArrowUpRight, Eyebrow, richTags } from "./Shared";
+import { ArrowUpRight, Eyebrow, richTags } from "./shared";
 
 export async function WritingList() {
   const t = await getTranslations("writing");
